analysis: add shared functions
We are going to be doing the same operations over,
many times, and this warrants a shared functions
file.
