+++ analysis/amg2023/data/flux-jobspec-and-events.json
@@ -27465,7 +27465,6 @@
     "aws/parallel-cluster/cpu/size32": [],
     "aws/parallel-cluster/cpu/size128": [],
     "aws/parallel-cluster/cpu/size256": [],
-    "aws/parallel-cluster/cpu/experiment": [],
     "azure/cyclecloud/cpu/size64": [],
     "azure/cyclecloud/cpu/size32": [],
     "azure/cyclecloud/cpu/size128": [],
@@ -27474,7 +27473,6 @@
     "azure/cyclecloud/gpu/size8": [],
     "azure/cyclecloud/gpu/size16": [],
     "azure/cyclecloud/gpu/size4": [],
-    "azure/aks/cpu/size64": [],
     "azure/aks/cpu/size32-placement": [
         {
             "jobspec": {
@@ -28787,9 +28785,6 @@
             ]
         }
     ],
-    "azure/aks/cpu/size32": [],
-    "azure/aks/cpu/size128": [],
-    "azure/aks/cpu/size256": [],
     "azure/aks/cpu/size64-placement": [
         {
             "jobspec": {
